requirements.txt
a2a-sdk==0.3.20
adk==0.0.5
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
akshare==1.17.91
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
Authlib==1.6.6
backoff==2.2.1
banks==2.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
black==25.11.0
brotli==1.2.0
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.5
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.12.0
cryptography==46.0.3
curl_cffi==0.13.0
dataclasses-json==0.6.7
ddgs==9.9.2
decorator==5.2.1
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
duckduckgo_search==8.1.1
durationpy==0.10
edgartools==4.34.1
et_xmlfile==2.0.0
fake-useragent==2.2.0
fastapi==0.123.5
fastuuid==0.14.0
filelock==3.20.0
filetype==1.2.0
flake8==7.3.0
flatbuffers==25.9.23
fredapi==0.5.2
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.12.0
google-adk==1.21.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.130.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-bigquery-storage==2.35.0
google-cloud-bigtable==2.34.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-logging==3.12.1
google-cloud-monitoring==2.28.0
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.60.0
google-cloud-speech==2.34.0
google-cloud-storage==3.7.0
google-cloud-trace==1.17.0
google-crc32c==1.7.1
google-genai==1.55.0
google-resumable-media==2.8.0
google_search_results==2.4.2
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.2.4
griffe==1.15.0
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.67.1
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hishel==0.1.3
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
httpxthrottlecache==0.3.0
huggingface-hub==0.36.0
humanfriendly==10.0
humanize==4.14.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpath==0.82.2
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-openai==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.12
langsmith==0.4.53
librt==0.6.3
litellm==1.80.9
llama-cloud==0.1.45
llama-cloud-services==0.6.87
llama-index-core==0.14.9
llama-index-instrumentation==0.4.2
llama-index-workflows==2.11.5
loguru==0.7.3
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mccabe==0.7.0
mcp==1.24.0
mdurl==0.1.2
mini-racer==0.12.4
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.12
mypy==1.19.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6
nltk==3.9.2
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.8.1
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.0
playwright==1.57.0
pluggy==1.6.0
posthog==5.4.0
praw==7.8.1
prawcore==2.4.0
primp==0.15.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.6
pyparsing==3.2.5
pypdf==6.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyrate-limiter==3.9.0
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
sec-edgar-downloader==5.0.3
sentry-sdk==2.47.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
SQLAlchemy==2.0.44
sqlalchemy-spanner==1.17.1
sqlparse==0.5.4
sse-starlette==3.0.3
stamina==25.1.0
starlette==0.50.0
steel==0.1
steel-sdk==0.15.0
structlog==25.5.0
sympy==1.14.0
tabulate==0.9.0
tavily==1.1.0
tenacity==9.1.2
textdistance==4.6.3
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.4.0
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.0.1
xlrd==2.0.2
xxhash==3.6.0
yarl==1.22.0
yfinance==0.2.66
zipp==3.23.0
zstandard==0.25.0
